requirements.txt
# Face Detection + Emotion Recognition System

## Required Packages

# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data science
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Logging and visualization
tensorboard>=2.13.0

# Optional: MTCNN face detection (more accurate but slower)
facenet-pytorch>=2.5.2

# Optional: Advanced augmentation
albumentations>=1.3.0
